+++ DCLearn/DCLearn/ContentView.swift
@@ -29,7 +29,7 @@ struct ContentView: View {
     var body: some View {
         NavigationView {
             ListView(options: listOptions, currentSelection: $currentOption)
-            
+
             switch currentOption {
             case .home:
                 MainView()
@@ -38,6 +38,8 @@ struct ContentView: View {
             }
         }
         .frame(minWidth: 600, minHeight: 400)
+        .navigationViewStyle(DoubleColumnNavigationViewStyle())
+        
     }
 }
 
@@ -45,57 +47,4 @@ struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
     }
-}
-
-
-struct MainView: View {
-    let columns: [GridItem] = [
-        .init(.flexible()), .init(.flexible()), .init(.flexible())
-    ]
-    let videoThumbnails = Video.samples
-    var body: some View {
-        VStack {
-            Image("header")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-            
-            LazyVGrid(columns: columns) {
-                ForEach(videoThumbnails) { item in
-                    VStack {
-                    Image(item.image)
-                        .resizable()
-                        .scaledToFill()
-                        .clipped()
-                        Text(item.title)
-                            .bold()
-                    }
-                }
-                
-            }
-//            Spacer()
-        }
-    }
-}
-
-
-struct Video: Identifiable {
-    let id = UUID()
-    let title: String
-    let description: String? = nil
-    let rating: Double = 0.0
-    let image: String
-    
-    
-    static let samples: [Video]  = [
-        .init(title: "SwiftUI", image: "swiftui"),
-        .init(title: "Kotlin", image: "kotlin"),
-        .init(title: "Front-End", image: "reactjs"),
-        .init(title: "Flutter", image: "flutter"),
-        .init(title: "ML Js", image: "mljs"),
-//        .init(title: "Node Js", image: "nodejs"),
-        .init(title: "Swift", image: "swift"),
-//        .init(title: "Web Dev", image: "web")
-        
-        
-    ]
 }
